+++ daemon/volumes_unix.go
@@ -9,6 +9,7 @@ import (
 	"strconv"
 	"strings"
 
+	mounttypes "github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/container"
 	"github.com/docker/docker/pkg/fileutils"
 	"github.com/docker/docker/pkg/mount"
@@ -57,6 +58,9 @@ func (daemon *Daemon) setupMounts(c *container.Container) ([]container.Mount, er
 				Destination: m.Destination,
 				Writable:    m.RW,
 				Propagation: string(m.Propagation),
+			}
+			if m.Spec.Type == mounttypes.TypeBind && m.Spec.BindOptions != nil {
+				mnt.NonRecursive = m.Spec.BindOptions.NonRecursive
 			}
 			if m.Volume != nil {
 				attributes := map[string]string{
@@ -129,11 +133,15 @@ func (daemon *Daemon) mountVolumes(container *container.Container) error {
 			return err
 		}
 
-		opts := "rbind,ro"
-		if m.Writable {
-			opts = "rbind,rw"
+		bindMode := "rbind"
+		if m.NonRecursive {
+			bindMode = "bind"
 		}
-
+		writeMode := "ro"
+		if m.Writable {
+			writeMode = "rw"
+		}
+		opts := strings.Join([]string{bindMode, writeMode}, ",")
 		if err := mount.Mount(m.Source, dest, bindMountType, opts); err != nil {
 			return err
 		}