+++ integration/internal/container/ops.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	containertypes "github.com/docker/docker/api/types/container"
+	mounttypes "github.com/docker/docker/api/types/mount"
 	networktypes "github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/api/types/strslice"
 	"github.com/docker/go-connections/nat"
@@ -65,6 +66,13 @@ func WithTty(tty bool) func(*TestContainerConfig) {
 func WithWorkingDir(dir string) func(*TestContainerConfig) {
 	return func(c *TestContainerConfig) {
 		c.Config.WorkingDir = dir
+	}
+}
+
+// WithMount adds an mount
+func WithMount(m mounttypes.Mount) func(*TestContainerConfig) {
+	return func(c *TestContainerConfig) {
+		c.HostConfig.Mounts = append(c.HostConfig.Mounts, m)
 	}
 }